maram autist/chatbot/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
transformers==4.18.0
torch==2.6.0
numpy==1.22.3
python-dotenv==0.19.2
nltk==3.6.5